# Conflicts:
#   frontend/src/Networking/Authentication/StoredTokenRepository.js
#   frontend/src/components/Scenes/Authentication/LogIn.vue

+++ backend/src/main/java/com/hva/ewa/team2/backend/domain/usecases/project/ProjectInteractor.java
@@ -1,7 +1,6 @@
 package com.hva.ewa.team2.backend.domain.usecases.project;
 
 import com.hva.ewa.team2.backend.common.services.asset.AssetService;
-import com.hva.ewa.team2.backend.common.services.date.DateServiceLogic;
 import com.hva.ewa.team2.backend.data.hourregistration.HourRegistrationRepository;
 import com.hva.ewa.team2.backend.data.project.ProjectRepository;
 import com.hva.ewa.team2.backend.data.user.UserRepository;
@@ -9,10 +8,12 @@ import com.hva.ewa.team2.backend.domain.models.project.Project;
 import com.hva.ewa.team2.backend.domain.models.project.ProjectFilter;
 import com.hva.ewa.team2.backend.domain.models.project.ProjectParticipant;
 import com.hva.ewa.team2.backend.domain.models.project.ProjectReport;
+import com.hva.ewa.team2.backend.domain.models.user.Admin;
 import com.hva.ewa.team2.backend.domain.models.user.Client;
 import com.hva.ewa.team2.backend.domain.models.user.Specialist;
 import com.hva.ewa.team2.backend.domain.models.user.User;
 import com.hva.ewa.team2.backend.rest.asset.json.FileResult;
+import com.hva.ewa.team2.backend.rest.exceptions.UnauthorizedException;
 import com.hva.ewa.team2.backend.rest.project.request.ProjectEditVerificationRequest;
 import com.hva.ewa.team2.backend.rest.project.request.ProjectInfoRequest;
 import com.hva.ewa.team2.backend.rest.project.request.ProjectParticipantAddInfoRequest;
@@ -26,9 +27,9 @@ import java.io.IOException;
 import java.time.DayOfWeek;
 import java.time.LocalDateTime;
 import java.time.temporal.TemporalAdjusters;
-import java.time.temporal.TemporalField;
-import java.time.temporal.WeekFields;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 
 @Component
 public class ProjectInteractor implements ProjectBusinessLogic {
@@ -36,24 +37,26 @@ public class ProjectInteractor implements ProjectBusinessLogic {
     private final UserRepository userRepo;
     private final ProjectRepository projectRepo;
     private final HourRegistrationRepository hourRegistrationRepo;
-    private final DateServiceLogic dateService;
     private final AssetService assetService;
 
     @Autowired
     public ProjectInteractor(ProjectRepository projectRepository,
                              UserRepository userRepo,
                              HourRegistrationRepository hourRegistrationRepo,
-                             DateServiceLogic dateService,
                              AssetService assetService) {
         this.projectRepo = projectRepository;
         this.userRepo = userRepo;
         this.hourRegistrationRepo = hourRegistrationRepo;
-        this.dateService = dateService;
         this.assetService = assetService;
     }
 
     @Override
-    public Project createProject(ProjectInfoRequest projectInfo) throws IOException {
+    public Project createProject(ProjectInfoRequest projectInfo, Integer userId) throws IOException {
+        final Optional<User> byId = userRepo.findById(userId);
+        if (!(byId.orElse(null) instanceof Admin)) {
+            throw new UnauthorizedException();
+        }
+
         final String title = projectInfo.getTitle();
         final Optional<Integer> clientId = projectInfo.getClient();
         final String description = projectInfo.getDescription();
@@ -84,21 +87,27 @@ public class ProjectInteractor implements ProjectBusinessLogic {
     }
 
     @Override
-    public Project getProjectInformation(int id) {
-        return getProjectOrThrowError(id);
+    public Project getProjectInformation(int id, Integer userId) {
+        User user = this.userRepo.findById(userId).orElse(null);
+        return getProjectOrThrowError(id, user, false);
     }
 
     @Override
-    public boolean deleteProject(int id) {
+    public boolean deleteProject(int id, Integer userId) {
         // throws error if project does not exist, therefore does not delete it.
-        getProjectOrThrowError(id);
+        User user = this.userRepo.findById(userId).orElse(null);
+        getProjectOrThrowError(id, user, true);
+
+        if (!(user instanceof Admin)) {
+            throw new UnauthorizedException();
+        }
 
         projectRepo.deleteById(id);
         return true;
     }
 
     @Override
-    public Project updateProjectInformation(int pId, ProjectInfoRequest body) throws IOException {
+    public Project updateProjectInformation(int pId, ProjectInfoRequest body, Integer userId) throws IOException {
         final String title = body.getTitle();
         final String description = body.getDescription();
         final MultipartFile logoUpload = body.getLogoFile();
@@ -108,7 +117,8 @@ public class ProjectInteractor implements ProjectBusinessLogic {
                 description
         );
 
-        final Project project = getProjectOrThrowError(pId);
+        User user = this.userRepo.findById(userId).orElse(null);
+        final Project project = getProjectOrThrowError(pId, user, true);
         project.setTitle(title);
         project.setDescription(description);
 
@@ -123,8 +133,9 @@ public class ProjectInteractor implements ProjectBusinessLogic {
     }
 
     @Override
-    public Project archiveProject(int id, ProjectEditVerificationRequest body, boolean unarchive) {
-        Project project = getProjectOrThrowError(id);
+    public Project archiveProject(int id, ProjectEditVerificationRequest body, boolean unarchive, Integer userId) {
+        User user = this.userRepo.findById(userId).orElse(null);
+        Project project = getProjectOrThrowError(id, user, true);
         String archiveWord = unarchive ? "unarchive" : "archive";
 
         if (!project.getTitle().equalsIgnoreCase(body.getTitle())) {
@@ -136,12 +147,17 @@ public class ProjectInteractor implements ProjectBusinessLogic {
         }
 
         project.setArchived(!unarchive);
-        return project;
+        return projectRepo.save(project);
     }
 
     @Override
-    public Project transferOwnership(int id, ProjectTransferRequest body) {
-        Project project = getProjectOrThrowError(id);
+    public Project transferOwnership(int id, ProjectTransferRequest body, Integer userId) {
+        User user = this.userRepo.findById(userId).orElse(null);
+        Project project = getProjectOrThrowError(id, user, true);
+
+        if (!(user instanceof Admin)) {
+            throw new UnauthorizedException();
+        }
 
         if (!project.getTitle().equalsIgnoreCase(body.getTitle())) {
             throw new IllegalArgumentException("Confirmation title does not match the project title.");
@@ -155,18 +171,20 @@ public class ProjectInteractor implements ProjectBusinessLogic {
         final Client client = validateClient(clientId);
         project.setClient(client);
 
-        return project;
+        return projectRepo.save(project);
     }
 
     @Override
-    public List<ProjectParticipant> getProjectParticipants(int id) {
-        final Project project = getProjectOrThrowError(id);
+    public List<ProjectParticipant> getProjectParticipants(int id, Integer userId) {
+        User user = this.userRepo.findById(userId).orElse(null);
+        final Project project = getProjectOrThrowError(id, user, false);
         return project.getParticipants();
     }
 
     @Override
-    public ProjectParticipant getProjectParticipant(int projectId, int userId) {
-        final Project project = getProjectOrThrowError(projectId);
+    public ProjectParticipant getProjectParticipant(int projectId, Integer requesterId, int userId) {
+        User user = this.userRepo.findById(requesterId).orElse(null);
+        final Project project = getProjectOrThrowError(projectId, user, false);
 
         final ProjectParticipant participant = project.getParticipantByUserId(userId);
         if (participant == null) {
@@ -177,8 +195,9 @@ public class ProjectInteractor implements ProjectBusinessLogic {
     }
 
     @Override
-    public ProjectParticipant removeProjectParticipant(int projectId, int userId) {
-        final Project project = getProjectOrThrowError(projectId);
+    public ProjectParticipant removeProjectParticipant(int projectId, Integer requesterId, int userId) {
+        User user = this.userRepo.findById(requesterId).orElse(null);
+        final Project project = getProjectOrThrowError(projectId, user, true);
 
         final ProjectParticipant participant = project.getParticipantByUserId(userId);
         if (participant == null) {
@@ -190,8 +209,9 @@ public class ProjectInteractor implements ProjectBusinessLogic {
     }
 
     @Override
-    public ProjectParticipant addProjectParticipant(int projectId, ProjectParticipantAddInfoRequest body) {
-        final Project project = getProjectOrThrowError(projectId);
+    public ProjectParticipant addProjectParticipant(int projectId, Integer requesterId, ProjectParticipantAddInfoRequest body) {
+        User user = this.userRepo.findById(requesterId).orElse(null);
+        final Project project = getProjectOrThrowError(projectId, user, true);
         final String role = body.getRole();
         if (role.isBlank()) {
             throw new IllegalArgumentException("The provided role is invalid (not provided/empty).");
@@ -215,83 +235,68 @@ public class ProjectInteractor implements ProjectBusinessLogic {
     }
 
     @Override
-    public List<ProjectReport> getProjectReports(int projectId, int userId) {
-        final Project project = getProjectOrThrowError(projectId);
-
-        // TODO: Integrate backend authorisation here.
-        final Optional<User> user = userRepo.findById(userId);
-
-        if (user.isEmpty()) {
-            throw new IllegalArgumentException("The user with ID " + userId + " does not exist.");
-        }
+    public List<ProjectReport> getProjectReports(int projectId, Integer userId) {
+        User user = this.userRepo.findById(userId).orElse(null);
+        getProjectOrThrowError(projectId, user, false);
 
         List<ProjectReport> reports = new ArrayList<>();
-
-        final TemporalField weekField = WeekFields.of(Locale.GERMANY).weekOfWeekBasedYear();
 
         LocalDateTime startMonth = LocalDateTime.now().with(TemporalAdjusters.firstDayOfMonth());
         LocalDateTime endMonth = LocalDateTime.now().with(TemporalAdjusters.lastDayOfMonth());
         LocalDateTime startWeek = LocalDateTime.now().with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
         LocalDateTime endWeek = LocalDateTime.now().with(TemporalAdjusters.nextOrSame(DayOfWeek.SUNDAY));
 
-        if (user.get() instanceof Specialist specialist) {
-            reports.add(new ProjectReport(
-                    "Totaal gemaakte uren",
-                    hourRegistrationRepo.getTotalHoursForProject(projectId, userId) + ""
-            ));
+        final boolean isSpecialist = user instanceof Specialist;
+        reports.add(new ProjectReport(
+                "Totaal gemaakte uren",
+                isSpecialist ? hourRegistrationRepo.getTotalHoursForProject(projectId, userId) + "" :
+                        hourRegistrationRepo.getTotalHoursForProject(projectId) + ""
+        ));
 
-            reports.add(new ProjectReport(
-                    "Gemaakte uren deze maand",
-                    hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, userId, startMonth, endMonth) + ""
-            ));
+        reports.add(new ProjectReport(
+                "Gemaakte uren deze maand",
+                isSpecialist ? hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, userId, startMonth, endMonth) + "" :
+                        hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, startMonth, endMonth) + ""
+        ));
 
-            reports.add(new ProjectReport(
-                    "Gemaakte uren deze week",
-                    hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, userId, startWeek, endWeek) + ""
-            ));
+        reports.add(new ProjectReport(
+                "Gemaakte uren deze week",
+                isSpecialist ? hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, userId, startWeek, endWeek) + "" :
+                        hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, startWeek, endWeek) + ""
+        ));
 
-            reports.add(new ProjectReport(
-                    "Verdiensten",
-                    String.format("€%,.2f", hourRegistrationRepo.getTotalRevenueForProject(projectId, userId)
-                    ).replace(".", ",")
-            ));
-        } else {
-            final Object totalHoursForProject = hourRegistrationRepo.getTotalHoursForProject(projectId);
-            System.out.println(totalHoursForProject);
-            reports.add(new ProjectReport(
-                    "Totaal gemaakte uren",
-                    totalHoursForProject + ""
-            ));
-
-            reports.add(new ProjectReport(
-                    "Gemaakte uren deze maand",
-                    hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, startMonth, endMonth) + ""
-            ));
-
-            reports.add(new ProjectReport(
-                    "Gemaakte uren deze week",
-                    hourRegistrationRepo.getTotalHoursForProjectBetween(projectId, startWeek, endWeek) + ""
-            ));
-
-            reports.add(new ProjectReport(
-                    "Ontwikkelkosten",
-                    String.format("€%.2f", hourRegistrationRepo.getTotalCostsForProject(projectId))
-                            .replace(".", ",")
-            ));
-        }
+        reports.add(new ProjectReport(
+                isSpecialist ? "Verdiensten" : "Ontwikkelkosten",
+                String.format("€%,.2f", isSpecialist ? hourRegistrationRepo.getTotalRevenueForProject(projectId, userId) :
+                        hourRegistrationRepo.getTotalCostsForProject(projectId)
+                ).replace(".", ",")
+        ));
 
         return reports;
     }
 
     @Override
-    public List<Project> getAllProjects(Optional<String> searchQuery, Optional<ProjectFilter> filter) {
-        if (searchQuery.isPresent() && filter.isPresent()) {
-            return projectRepo.findAll(filter.get(), searchQuery.get());
-        } else if (searchQuery.isPresent()) {
-            return projectRepo.findAll(searchQuery.get());
-        } else if (filter.isPresent()) {
-            return projectRepo.findAll(filter.get());
+    public List<Project> getAllProjects(Optional<String> searchQuery, Optional<String> filter, Integer userId) {
+        final User user = userRepo.findById(userId).orElse(null);
+        if (user == null) {
+            throw new UnauthorizedException();
+        }
+
+        boolean queryPresent = searchQuery.isPresent();
+        boolean filterPresent = filter.isPresent();
+        boolean isAdmin = user instanceof Admin;
+
+        if (queryPresent && filterPresent) {
+            if (!isAdmin) return projectRepo.findAllByQuery(ProjectFilter.valueOf(filter.get()), searchQuery.get(), userId);
+            return projectRepo.findAllByQuery(ProjectFilter.valueOf(filter.get()), searchQuery.get());
+        } else if (queryPresent) {
+            if (!isAdmin) return projectRepo.findAllByQuery(searchQuery.get(), userId);
+            return projectRepo.findAllByQuery(searchQuery.get());
+        } else if (filterPresent) {
+            if (!isAdmin) return projectRepo.findAll(ProjectFilter.valueOf(filter.get()), userId);
+            return projectRepo.findAll(ProjectFilter.valueOf(filter.get()));
         } else {
+            if (!isAdmin) return projectRepo.findAll(userId);
             return projectRepo.findAll();
         }
     }
@@ -312,12 +317,36 @@ public class ProjectInteractor implements ProjectBusinessLogic {
         return casted;
     }
 
-    private Project getProjectOrThrowError(int id) throws IllegalArgumentException {
+    public Project getProjectOrThrowError(int id, User user, boolean requiresModerationPerm) throws IllegalArgumentException {
         final Optional<Project> project = projectRepo.findById(id);
         if (project.isEmpty()) {
             throw new IllegalArgumentException("The project with ID " + id + " does not exist.");
         }
+        if (!project.get().hasAccess(user, requiresModerationPerm)) {
+            throw new UnauthorizedException();
+        }
+
         return project.get();
     }
 
+    @Override
+    public int getProjectCount(Integer userId) {
+        final User user = userRepo.findById(userId).orElse(null);
+        if (user == null) {
+            throw new UnauthorizedException();
+        }
+
+        if (user instanceof Admin) return projectRepo.getCount();
+        return projectRepo.getCount(userId);
+    }
+
+    @Override
+    public Double getEarnings(Integer userId) {
+        return hourRegistrationRepo.getTotalRevenueForUser(userId);
+    }
+
+    @Override
+    public Double getHours(Integer userId) {
+        return hourRegistrationRepo.getTotalHoursForUser(userId);
+    }
 }
